+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/base_stats/BaseStatsAdder.java
@@ -2,12 +2,15 @@ package com.robertx22.age_of_exile.aoe_data.database.base_stats;
 
 import com.robertx22.age_of_exile.aoe_data.database.stats.Stats;
 import com.robertx22.age_of_exile.database.data.base_stats.BaseStatsConfig;
+import com.robertx22.age_of_exile.database.data.stats.types.generated.AttackDamage;
+import com.robertx22.age_of_exile.database.data.stats.types.offense.SpellPower;
 import com.robertx22.age_of_exile.database.data.stats.types.resources.energy.Energy;
 import com.robertx22.age_of_exile.database.data.stats.types.resources.energy.EnergyRegen;
 import com.robertx22.age_of_exile.database.data.stats.types.resources.health.Health;
 import com.robertx22.age_of_exile.database.data.stats.types.resources.health.HealthRegen;
 import com.robertx22.age_of_exile.database.data.stats.types.resources.mana.Mana;
 import com.robertx22.age_of_exile.database.data.stats.types.resources.mana.ManaRegen;
+import com.robertx22.age_of_exile.uncommon.enumclasses.Elements;
 import com.robertx22.library_of_exile.registry.ExileRegistryInit;
 
 public class BaseStatsAdder implements ExileRegistryInit {
@@ -26,6 +29,9 @@ public class BaseStatsAdder implements ExileRegistryInit {
         c.id = PLAYER;
 
         c.nonScaled(Mana.getInstance(), 10);
+
+        c.scaled(SpellPower.getInstance(), 1);
+        c.scaled(new AttackDamage(Elements.Physical), 1);
 
         c.scaled(Health.getInstance(), 10);
         c.scaled(Mana.getInstance(), 20);