+++ src/main/java/com/robertx22/age_of_exile/aoe_data/database/stats/old/DatapackStats.java
@@ -33,16 +33,16 @@ import java.util.UUID;
 public class DatapackStats implements ExileRegistryInit {
 
     public static Stat INT = new CoreStat("intelligence", "Intelligence", CoreStatData.of(Arrays.asList(
-        new OptScaleExactStat(0.2F, SpellPower.getInstance(), ModType.FLAT),
-        new OptScaleExactStat(2, new ElementalResist(Elements.Elemental), ModType.FLAT)
+        new OptScaleExactStat(0.1F, SpellPower.getInstance(), ModType.FLAT),
+        new OptScaleExactStat(1, new ElementalResist(Elements.Elemental), ModType.FLAT)
     )));
     public static Stat STR = new CoreStat("strength", "Strength", CoreStatData.of(Arrays.asList(
-        new OptScaleExactStat(0.2F, new AttackDamage(Elements.Physical), ModType.FLAT),
-        new OptScaleExactStat(2, Armor.getInstance(), ModType.FLAT)
+        new OptScaleExactStat(0.1F, new AttackDamage(Elements.Physical), ModType.FLAT),
+        new OptScaleExactStat(1, Armor.getInstance(), ModType.FLAT)
     )));
     public static Stat DEX = new CoreStat("dexterity", "Dexterity", CoreStatData.of(Arrays.asList(
-        new OptScaleExactStat(0.2F, new AttackDamage(Elements.Physical), ModType.FLAT),
-        new OptScaleExactStat(2, DodgeRating.getInstance(), ModType.FLAT)
+        new OptScaleExactStat(0.1F, new AttackDamage(Elements.Physical), ModType.FLAT),
+        new OptScaleExactStat(1, DodgeRating.getInstance(), ModType.FLAT)
     )));
     public static Stat VIT = new CoreStat("vitality", "Vitality", CoreStatData.of(Arrays.asList(
         new OptScaleExactStat(2, Health.getInstance(), ModType.FLAT),