+++ src/main/java/com/robertx22/age_of_exile/player_skills/ingredient/items/IngredientItem.java
@@ -38,10 +38,6 @@ public class IngredientItem extends AutoItem {
 
                 for (int tier = 0; tier <= LevelUtils.getMaxTier(); tier++) {
 
-                    int test = LevelUtils.getMaxTier();
-
-                    System.out.println(test);
-
                     IngredientData data = new IngredientData();
                     data.tier = tier;
                     data.id = ing.GUID();