+++ src/main/java/com/robertx22/age_of_exile/uncommon/utilityclasses/GearSoulOnInvTick.java
@@ -47,7 +47,7 @@ public class GearSoulOnInvTick {
 
             if (soul != null && soul.canInsertIntoStack(stack)) {
                 try {
-                    GearItemData gear = soul.createGearData();
+                    GearItemData gear = soul.createGearData(stack);
                     stack.removeTagKey(StatSoulItem.TAG);
                     gear.saveToStack(stack);
                 } catch (Exception e) {