+++ src/main/java/com/robertx22/age_of_exile/vanilla_mc/commands/CommandRefs.java
@@ -1,6 +1,6 @@
 package com.robertx22.age_of_exile.vanilla_mc.commands;
 
 public class CommandRefs {
-    public static String ID = "age_of_exile";
+    public static String ID = "mine_and_slash";
 
 }