+++ src/main/java/com/robertx22/age_of_exile/vanilla_mc/items/gemrunes/BaseGemRuneItem.java
@@ -5,6 +5,7 @@ import com.robertx22.age_of_exile.database.data.BaseRuneGem;
 import com.robertx22.age_of_exile.database.data.StatModifier;
 import com.robertx22.age_of_exile.database.data.gear_types.bases.SlotFamily;
 import com.robertx22.age_of_exile.database.registry.ExileDB;
+import com.robertx22.age_of_exile.saveclasses.gearitem.gear_bases.TooltipInfo;
 import com.robertx22.age_of_exile.uncommon.datasaving.Load;
 import com.robertx22.age_of_exile.uncommon.utilityclasses.ClientOnly;
 import com.robertx22.age_of_exile.uncommon.utilityclasses.TooltipUtils;
@@ -60,25 +61,27 @@ public abstract class BaseGemRuneItem extends Item {
         int efflvl = Load.Unit(ClientOnly.getPlayer())
             .getLevel();
 
+        TooltipInfo info = new TooltipInfo();
+
         tooltip.add(new StringTextComponent(""));
-        List<StatModifier> wep = gem.getFor(SlotFamily.Weapon);
+        List<OptScaleExactStat> wep = gem.getFor(SlotFamily.Weapon);
         tooltip.add(new StringTextComponent("On Weapon:").withStyle(TextFormatting.RED));
-        for (StatModifier x : wep) {
-            tooltip.addAll(x.getEstimationTooltip(efflvl));
+        for (OptScaleExactStat x : wep) {
+            tooltip.addAll(x.GetTooltipString(info));
         }
 
         tooltip.add(new StringTextComponent(""));
-        List<StatModifier> armor = gem.getFor(SlotFamily.Armor);
+        List<OptScaleExactStat> armor = gem.getFor(SlotFamily.Armor);
         tooltip.add(new StringTextComponent("On Armor:").withStyle(TextFormatting.BLUE));
-        for (StatModifier x : armor) {
-            tooltip.addAll(x.getEstimationTooltip(efflvl));
+        for (OptScaleExactStat x : armor) {
+            tooltip.addAll(x.GetTooltipString(info));
         }
 
         tooltip.add(new StringTextComponent(""));
-        List<StatModifier> jewelry = gem.getFor(SlotFamily.Jewelry);
+        List<OptScaleExactStat> jewelry = gem.getFor(SlotFamily.Jewelry);
         tooltip.add(new StringTextComponent("On Jewelry:").withStyle(TextFormatting.LIGHT_PURPLE));
-        for (StatModifier x : jewelry) {
-            tooltip.addAll(x.getEstimationTooltip(efflvl));
+        for (OptScaleExactStat x : jewelry) {
+            tooltip.addAll(x.GetTooltipString(info));
         }
 
         tooltip.add(new StringTextComponent(""));